+++ CHANGELOG.md
@@ -54,6 +54,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Protein and PolypeptideChain getSequence methods to use new sequence package
 - All structure classes are now transformable.
 - Atom constructors now take doubles for the coordinates rather than Strings.
+- In Metrics, changed the score calculation to be based on the number of
+  residues in the reference structure rather than the number of Residues
+  in the alignment. This makes the scores more stable and makes more
+  sense when evaluating models. If a model only includes a fraction of the
+  residues that are in the reference, it should be penalized for the
+  missing residues. 
 
 ## [0.6.4] 2017-10-27
 ### Fixed